+++ crates/cli/src/commands/graph/mod.rs
@@ -1,3 +1,4 @@
+pub mod action;
 pub mod dep;
 mod dto;
 pub mod project;
